perf: Prefetch references on hover
